+++ client/src/router/index.js
@@ -38,7 +38,7 @@ import TokenStorage from '@/lib/TokenStorage.js'
 // checkDesigner
 const checkDesigner=(to, from, next) => {
   var payload = TokenStorage.getPayload()
-  if(payload?.user?.options?.showDesigner ?? payload?.user?.roles?.includes("admin")){
+  if(payload?.user?.options?.showDesigner){
     next()
   }else{
     console.log("You don't have access to the designer")
@@ -47,7 +47,7 @@ const checkDesigner=(to, from, next) => {
 // checkLogs
 const checkLogs=(to, from, next) => {
   var payload = TokenStorage.getPayload()
-  if(payload?.user?.options?.showLogs ?? payload?.user?.roles?.includes("admin")){
+  if(payload?.user?.options?.showLogs){
     next()
   }else{
     console.log("You don't have access to the logs")
@@ -56,7 +56,7 @@ const checkLogs=(to, from, next) => {
 // checkJobs
 const checkJobs=(to, from, next) => {
   var payload = TokenStorage.getPayload()
-  if(payload?.user?.options?.showJobs ?? true){
+  if(payload?.user?.options?.showJobs){
     next()
   }else{
     console.log("You don't have access to the jobs")
@@ -65,7 +65,7 @@ const checkJobs=(to, from, next) => {
 // checkSettings
 const checkSettings=(to, from, next) => {
   var payload = TokenStorage.getPayload()
-  if(payload?.user?.options?.showSettings ?? payload?.user?.roles?.includes("admin")){
+  if(payload?.user?.options?.showSettings){
     next()
   }else{
     console.log("You don't have access to the settings")
@@ -74,7 +74,7 @@ const checkSettings=(to, from, next) => {
 // allowBackupOps
 const allowBackupOps=(to, from, next) => {
   var payload = TokenStorage.getPayload()
-  if(payload?.user?.options?.allowBackupOps ?? payload?.user?.roles?.includes("admin")){
+  if(payload?.user?.options?.allowBackupOps){
     next()
   }else{
     console.log("You don't have access to the backups page")